Check that basedir is defined
git-svn-id: https://svn.apache.org/repos/asf/maven/wagon/trunk@410414 13f79535-47bb-0310-9956-ffa450edef68

+++ wagon-providers/wagon-webdav/src/test/java/org/apache/maven/wagon/providers/webdav/AuthenticatedWebDavWagonTest.java
@@ -58,6 +58,11 @@ public class AuthenticatedWebDavWagonTest
     protected void setupWagonTestingFixtures()
         throws Exception
     {
+        if ( System.getProperty( "basedir" ) == null )
+        {
+            fail( "System property 'basedir' must be set for the web server to run properly" );
+        }
+
         File file = FileTestUtils.createUniqueFile( "authdav-repository", "test-resource" );
 
         file.delete();
